requirements.txt
torch          
torchaudio

numpy==1.24.0
scipy>=1.7.0
librosa>=0.10.0
einops>=0.6.0
julius>=0.2.6
lameenc>=1.4.2
openunmix>=1.2.1
hydra-core>=1.1.0
hydra-colorlog>=1.1.0
flashy>=0.0.1
retrying>=1.3.3

noisereduce>=2.0.1
sounddevice>=0.4.6
omegaconf>=2.2.2

submitit>=1.4.0
cloudpickle>=2.0.0
treetable>=0.2.0

# pip install git+https://github.com/adefossez/demucs.git --no-deps
# pip install git+https://github.com/facebookresearch/dora.git --no-deps

onnx
onnxruntime

# FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Redis client
redis>=5.0.0

# Environment management
python-dotenv>=1.0.0